requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
babel==2.16.0
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
Brotli==1.0.9
build==1.2.2.post1
CacheControl==0.12.11
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
coverage==5.3.1
crashtest==0.4.1
cryptography==43.0.3
cssselect2==0.7.0
cycler==0.12.1
data-gradients==0.3.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.7.0
docutils==0.17.1
dulwich==0.21.3
einops==0.3.2
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
grpcio==1.67.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
ifaddr==0.2.0
imagededup==0.3.2
imageio==2.36.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.6.0
jaraco.classes==3.2.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json-tricks==3.16.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.13.1
kiwisolver==1.4.7
lazy_loader==0.4
lockfile==0.12.2
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
narwhals==1.10.0
networkx==3.2.1
numpy==1.23.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.0
onnxsim==0.4.36
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
pexpect==4.8.0
pillow==10.4.0
pip==24.2
pip-tools==7.4.1
pkginfo==1.10.0
platformdirs==2.5.2
pluggy==1.5.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.0
pydeck==0.9.1
pyDeprecate==0.3.2
pyee==12.0.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.0.1
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.3.3
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
qrcode==8.0
qudida==0.0.4
rapidfuzz==3.5.2
referencing==0.35.1
reportlab==3.6.13
requests==2.32.3
requests-toolbelt==0.9.1
rich==13.9.3
rpds-py==0.20.0
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.37.2
streamlit==1.39.0
streamlit-webrtc==0.47.9
stringcase==1.2.0
super-gradients==3.7.1
supervision==0.21.0
svglib==1.5.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.5.0
torchmetrics==0.8.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
treelib==1.6.1
triton==3.1.0
trove-classifiers==2024.10.18
twilio==9.3.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.17.1
watchdog==5.0.3
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
xhtml2pdf==0.2.11
XlsxWriter==3.2.0
yarl==1.16.0
zipp==3.20.2
